requirements.txt
boto3==1.35.21
PyYAML==6.0.2
pandas==2.2.2
